requirements.txt
streamlit
openai-whisper
PyPDF2
langchain
python-dotenv
moviepy
faiss-cpu
langchain-groq
langchain_community
moviepy
pytube
openai